fix logics of updating total_amount

+++ db/migrate/20220906093800_create_table_customers.rb
@@ -4,9 +4,9 @@ class CreateTableCustomers < ActiveRecord::Migration[7.0]
       t.string :name
       t.boolean :is_captain
       t.boolean :is_paid
-      t.integer :amount_due
-      t.integer :tip_amount
-      t.integer :total_amount
+      t.float :amount_due
+      t.float :tip_amount
+      t.float :total_amount
 
       t.timestamps
     end
